Only the last commit is new. These numbers are for *just that commit*.

```
benchmark                                 old ns/op     new ns/op     delta
BenchmarkFieldSet/serialize-20-12         27360         15191         -44.48%
BenchmarkFieldSet/deserialize-20-12       35718         35840         +0.34%
BenchmarkFieldSet/serialize-50-12         57494         37869         -34.13%
BenchmarkFieldSet/deserialize-50-12       86494         82157         -5.01%
BenchmarkFieldSet/serialize-100-12        170144        111428        -34.51%
BenchmarkFieldSet/deserialize-100-12      223283        224378        +0.49%
BenchmarkFieldSet/serialize-500-12        919350        597742        -34.98%
BenchmarkFieldSet/deserialize-500-12      1084716       1062631       -2.04%
BenchmarkFieldSet/serialize-1000-12       1786525       1167377       -34.66%
BenchmarkFieldSet/deserialize-1000-12     1950813       2292026       +17.49%

benchmark                                 old allocs     new allocs     delta
BenchmarkFieldSet/serialize-20-12         110            25             -77.27%
BenchmarkFieldSet/deserialize-20-12       216            216            +0.00%
BenchmarkFieldSet/serialize-50-12         300            50             -83.33%
BenchmarkFieldSet/deserialize-50-12       616            614            -0.32%
BenchmarkFieldSet/serialize-100-12        1032           95             -90.79%
BenchmarkFieldSet/deserialize-100-12      2167           2166           -0.05%
BenchmarkFieldSet/serialize-500-12        5701           472            -91.72%
BenchmarkFieldSet/deserialize-500-12      10770          10779          +0.08%
BenchmarkFieldSet/serialize-1000-12       12365          1000           -91.91%
BenchmarkFieldSet/deserialize-1000-12     23523          23534          +0.05%
```

Unfortunately I don't think I can give deserialize the same treatment without making jsoniter changes :/

Note that this makes use of the unsafe package, I need changes to jsoniter to avoid that.

Jsoniter's pools are apparently... not for the purpose of doing optimal memory management.