@lavalamp 2019-07-24
This brings the total improvement to:

```
benchmark                       old ns/op     new ns/op     delta
BenchmarkApplyNewObject-12      2230683       1933055       -13.34%
BenchmarkUpdateNewObject-12     2624417       2236517       -14.78%
BenchmarkRepeatedUpdate-12      1031653       755754        -26.74%
BenchmarkSetToFields-12         24778         5078          -79.51%
BenchmarkFieldsToSet-12         33876         10309         -69.57%
BenchmarkYAMLToTyped-12         68377         65714         -3.89%

benchmark                       old allocs     new allocs     delta
BenchmarkApplyNewObject-12      8632           6827           -20.91%
BenchmarkUpdateNewObject-12     11210          7913           -29.41%
BenchmarkRepeatedUpdate-12      3057           2288           -25.16%
BenchmarkSetToFields-12         124            15             -87.90%
BenchmarkFieldsToSet-12         147            82             -44.22%
BenchmarkYAMLToTyped-12         150            152            +1.33%
```