@lavalamp 2019-07-24
The unsafe bit is key, sadly-- removing it I get

```
benchmark                                 old allocs     new allocs     delta
BenchmarkFieldSet/serialize-20-12         25             96             +284.00%
BenchmarkFieldSet/deserialize-20-12       216            216            +0.00%
BenchmarkFieldSet/serialize-50-12         50             254            +408.00%
BenchmarkFieldSet/deserialize-50-12       614            616            +0.33%
BenchmarkFieldSet/serialize-100-12        95             803            +745.26%
BenchmarkFieldSet/deserialize-100-12      2166           2168           +0.09%
BenchmarkFieldSet/serialize-500-12        472            4013           +750.21%
BenchmarkFieldSet/deserialize-500-12      10779          10773          -0.06%
BenchmarkFieldSet/serialize-1000-12       1000           8654           +765.40%
BenchmarkFieldSet/deserialize-1000-12     23534          23511          -0.10%
```