@lavalamp 2019-07-24
https://github.com/json-iterator/go/pull/393