@lavalamp 2019-07-24
I've modified this to do a less unsafe thing with the unsafe package.