@apelisse 2019-07-24
I think this is OK. I don't have much knowledge about the jsoniter though